+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ## 2021.2
 
 * I switched to the .NET 5.0 framework
+* bug fixes
+* I added TryAdd overloads 
 
 ## 2021.1